Fix width of progress bars
The progress package does not take into account colors when calculating
the width, so we need to increase our requested progress bar width by
the width of the ansii escapes.

We also needed to change how progress bars work with the threaded
connection reading. Reading from connections is not thread safe and must
happen from the main R thread.