allow %d and %m to take either 1 or 2 characters, e.g. 03/04/199… (#160)
* First shot, but still no solution to #123

The problem seems that dates like 1/5/1999 are not read with
format "%m/%d/%Y", because 01/05/1999 is expected. I think
that if you read ?strptime well, this is correct, but then base
R works differently and accepts single digit days and months. So,
I think the preferred behavior is to accept them.

This PR is messy and full with print statements. I believe that the
line marked with "need to set back the parser iterator/pointer one
character here?" should be replaced with setting back the parser
pointer one step, as we have a fail read here which does NOT raise
an error. The error then happens in next step, when the "/" (or other
separator) has been absorbed, I suspect. Happy to revisit this when
I know how the parser works.

* allow %d and %m to be 1 or 2 characters long; fix #123

* tidy

* add NEWS entry

Fixes #123
Fixes #170