Fix n_max behavior for windows newlines from connections
Previously the index was getting spurious entries for the size of the
whole file and an extra newline, which caused the row number calculation
to be off.

The test is complicated by the fact that this behavior is only observed
if the file does not fit into a single read buffer.

Fixes #244