Enable factor level encoding test
This was tricky because the string literal input in vroom always converts itself
to UTF-8, so we need to write a file as latin-1 instead. Writing to
latin-1 itself is tricky because nearly every base R function internally
converts strings to the native locale before writing. SO we need to be
very explicit in order to write the correct bytes.