Handle case when windows newlines are in separate connection buffers
We would previously inadvertently index these cases twice, once for the
'\r' and once for the '\n'. This caused difficult to track down bugs
when reading files with windows line endings from connections, such as
compressed files. The file would read fine until a line ending happened
to fall on the connection buffer boundary, then the rest of the file
would have a garbled index.

Fixes #331