Fix performance issue with subset iterators
Previously we were creating a new iterator each time we returned a
value, now we store the previous index in the iterator and update the
full iterator with the difference when needed.

The performance issue was occurring due to transient object creation of
the temporary iterators and moving them to the right location.

Fixes #378