Support '%h' format for time durations
This was added to readr after the datetime code was copied to vroom.

This ports the code to the current vroom codebase.

It also fixes a bug where thrown exceptions were not properly caught in
`vroom_time()`, which caused R to crash if an invalid time format was
supplied.

Fixes https://github.com/tidyverse/readr/issues/1312