Ensure fixed width files that do not end with a newline are properly parsed
We need to make sure we add a newline at the end of the file in these
cases. An additional complication addressed by this code is dealing with
the case when a CRLF spans the buffer between two files. We solve this
by only looking for either '\n' or '\r' character, but never both, so we
don't double count the same newline.

Fixes https://github.com/tidyverse/readr/issues/1293