Clear warnings before restarting parsing for newline errors
When we encounter an embedded newline in a location we weren't expecting
we need to restart parsing without multiple threads. In this case we
could have already collected some (false) problems during the
multi-threaded reading, and we need to clear those when we restart
parsing.

Fixes https://github.com/tidyverse/readr/issues/1313