Fix indexing subsets when the index hits the file boundary
Unfortunately the iterators as currently designed don't really support
both positive and negative traversal. This is because we use an unsigned
`size_t` type and in order to support negative traversal we would need
to have a reverse end that was 1 less than the start, not really
possible with the current design.

So instead we remove the broken `prev()` API entirely and if we need to
use previous indexes we recreate a new iterator from the start.

This could have bad performance implications if you index from
end:start, but that will hopefully be rare in practice.

Fixes #383