Don't dereference begin when begin == end == end of file
Fixes #429

As described in the issue, this file lacks a newline at the end AND is missing the last field entirely. It also uses CRLF line endings.

So the entries in the index for the beginning and end of the last field are essentially degenerate, i.e. they are equal to each other and at/past the end of the file.

The segfault appears reliably in RStudio. Less so in VS Code, with a debugger attached, perhaps with breakpoints in place. A Heisenbug.

But if end <= begin, it's clear we don't need to worry about un-quoting, so I've added a check for begin < end before we enter the branch where we will dereference begin.