Fix `subset_iterator::index()` to explicitly position wrapped iterator (#568)
* Fix subset_iterator::index() to explicitly position wrapped iterator

subset_iterator is different from column_iterator and row_iterator, in that it delegates to a wrapped iterator. This wrapped iterator is positioned lazily inside value(), but index() can be called independently (e.g., during error reporting). When index() is called before value(), there's no guarantee that it's positioned correctly and, if not, could return bad data.

 This fix calculates index() by explicitly advancing from start_ to the correct position, making subset_iterator work independently like the other iterators rather than assuming it_ is already positioned.

* Add a test

* Remove dead code

If we kept an implementation of position() here, then we'd need to answer the question of why it starts with `it_`, but we know that's not correct for index(). I prefer to remove it until there's concrete proof that this method is more than theoretical.

* Test with and without column names

* Missing include

* Different include

* Add NEWS bullet